# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.LoginResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
...
  models.UpdateUserStatusRequest:
    properties:
      is_active:
        type: boolean
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for VPN management system supporting SSH, VMESS, VLESS, Trojan,
    and Shadowsocks protocols
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: VPN API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate admin user and get JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Admin login
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new admin user account
      parameters:
      - description: Registration credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Invalid request or user already exists
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Register new admin user
      tags:
      - Authentication
...
  /user/{username}/status:
    put:
      consumes:
      - application/json
      description: Enable or disable a user account
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: User status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User status updated successfully
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Failed to update user status
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Update user status
      tags:
      - User Management
  /user/password:
    put:
      consumes:
      - application/json
      description: Change authenticated user's password
      parameters:
      - description: Old and new passwords
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - User Management
  /user/profile:
    get:
      consumes:
      - application/json
      description: Get authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: Profile retrieved successfully
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User Management
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"